Using JedisCluster client for redis cluster module
