+++ redis-persistence/src/main/java/com/netflix/conductor/jedis/RedisClusterJedisProvider.java
@@ -12,6 +12,7 @@
  */
 package com.netflix.conductor.jedis;
 
+import com.netflix.conductor.dyno.DynomiteConfiguration;
 import com.netflix.dyno.connectionpool.Host;
 import com.netflix.dyno.connectionpool.HostSupplier;
 import java.util.ArrayList;
@@ -20,37 +21,37 @@ import java.util.Set;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
+
+import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
 import redis.clients.jedis.HostAndPort;
 import redis.clients.jedis.commands.JedisCommands;
 
 public class RedisClusterJedisProvider implements Provider<JedisCommands> {
 
     private final HostSupplier hostSupplier;
+    private final DynomiteConfiguration configuration;
 
     @Inject
-    public RedisClusterJedisProvider(HostSupplier hostSupplier){
+    public RedisClusterJedisProvider(HostSupplier hostSupplier, DynomiteConfiguration configuration){
         this.hostSupplier = hostSupplier;
+        this.configuration = configuration;
     }
 
     @Override
     public JedisCommands get() {
-        // FIXME This doesn't seem very safe, but is how it was in the code this was moved from.
-//        Host host = new ArrayList<>(hostSupplier.getHosts()).get(0);
-//        GenericObjectPoolConfig poolConfig = new GenericObjectPoolConfig();
-//        poolConfig.setMinIdle(5);
-//        poolConfig.setMaxTotal(1000);
-//
-//        JedisPool jedisPool = new JedisPool(poolConfig, host.getHostName(), host.getPort());
-//        return new JedisCluster(jedisPool);
+
+        GenericObjectPoolConfig poolConfig = new GenericObjectPoolConfig();
+        poolConfig.setMinIdle(configuration.getMinIdleConnectionsPerHost());
+        poolConfig.setMaxTotal(configuration.getMaxConnectionsPerHost());
 
         Host host = new ArrayList<>(hostSupplier.getHosts()).get(0);
 
-        Set<HostAndPort> jedisClusterNodes = new HashSet<>();
         //Jedis Cluster will attempt to discover cluster nodes automatically
+        Set<HostAndPort> jedisClusterNodes = new HashSet<>();
         jedisClusterNodes.add(new HostAndPort(host.getHostName(), host.getPort()));
 
-        redis.clients.jedis.JedisCluster clusterClient = new redis.clients.jedis.JedisCluster(jedisClusterNodes);
+        redis.clients.jedis.JedisCluster clusterClient = new redis.clients.jedis.JedisCluster(jedisClusterNodes, poolConfig);
 
-        return new JedisClusterMod(clusterClient);
+        return new JedisCluster(clusterClient);
     }
 }
